refactor(datasets): extract metadata CRUD routes to routes/metadata.ts (6d)
list/get/schema/safe-schema/patch/change-owner/delete + the permissions sub-router,
as typed adapters using dataset/datasetFull/publicationSite/publicBaseUrl accessors.
Source fixes: permissions.canDoForOwnerMiddleware req->base Request; preparePatch
draftValidationMode made optional (the PATCH route only ever passes 5 args — parking lot).
The 6 dropped @ts-ignores in this block are gone; clean(req) cast to the #types Request
at the boundary; journals.log event cast as Event (cf. applications/service.ts).
