+++ api/src/datasets/router.js
@@ -1,7 +1,6 @@
 import express from 'express'
 import * as ajv from '../misc/utils/ajv.ts'
 import fs from 'fs-extra'
-import moment from 'moment'
 import { httpError } from '@data-fair/lib-utils/http-errors.js'
 import sanitizeHtml from '@data-fair/data-fair-shared/sanitize-html.js'
 import equal from 'deep-equal'
@@ -16,7 +15,7 @@ import privateDatasetAPIDocs from '../../contract/dataset-private-api-docs.ts'
 import * as permissions from '../misc/utils/permissions.ts'
 import * as usersUtils from '../misc/utils/users.ts'
 import * as datasetUtils from './utils/index.ts'
-import { updateStorage, updateTotalStorage } from './utils/storage.ts'
+import { updateStorage } from './utils/storage.ts'
 import * as restDatasetsUtils from './utils/rest.ts'
 import clone from '@data-fair/lib-utils/clone.js'
 import * as cacheHeaders from '../misc/utils/cache-headers.ts'
@@ -30,22 +29,20 @@ import * as publicationSites from '../misc/utils/publication-sites.ts'
 import * as clamav from '../misc/utils/clamav.ts'
 import * as rateLimiting from '../misc/utils/rate-limiting.ts'
 import { syncDataset as syncRemoteService } from '../remote-services/service.ts'
-import { findDatasets, applyPatch, deleteDataset, createDataset, memoizedGetDataset, cancelDraft } from './service.ts'
-import { tableSchema, jsonSchema, getSchemaBreakingChanges, filterSchema } from './utils/data-schema.ts'
-import { dir } from './utils/files.ts'
+import { applyPatch, createDataset, memoizedGetDataset, cancelDraft } from './service.ts'
 import { preparePatch } from './utils/patch.ts'
 import { checkStorage, lockDataset, readDataset } from './middlewares.ts'
-import { apiKeyMiddlewareRead, apiKeyMiddlewareWrite, apiKeyMiddlewareAdmin } from './routes/_common.ts'
+import { apiKeyMiddlewareRead, apiKeyMiddlewareWrite } from './routes/_common.ts'
 import { registerMasterDataRoutes } from './routes/master-data.ts'
 import { registerReadRoutes } from './routes/read.ts'
 import { registerLinesRoutes } from './routes/lines.ts'
 import { registerOwnLinesRoutes } from './routes/own-lines.ts'
 import { registerFilesRoutes } from './routes/files.ts'
+import { registerMetadataRoutes } from './routes/metadata.ts'
 import config from '#config'
 import mongo from '#mongo'
 import debugModule from 'debug'
-import contentDisposition from 'content-disposition'
-import { reqAdminMode, reqSession, reqSessionAuthenticated, session } from '@data-fair/lib-express'
+import { reqAdminMode, reqSession, reqSessionAuthenticated } from '@data-fair/lib-express'
 import eventsQueue from '@data-fair/lib-node/events-queue.js'
 import eventsLog from '@data-fair/lib-express/events-log.js'
 import { getFlatten } from './utils/flatten.ts'
@@ -67,285 +64,7 @@ router.use((req, res, next) => {
   next()
 })
 
-// Get the list of datasets
-router.get('', apiKeyMiddlewareRead, rateLimiting.middleware, cacheHeaders.listBased, async (req, res) => {
-  // @ts-ignore
-  const publicationSite = req.publicationSite
-  // @ts-ignore
-  const publicBaseUrl = req.publicBaseUrl
-  const reqQuery = /** @type {Record<string, string>} */(req.query)
-
-  const response = await findDatasets(mongo.db, req.getLocale(), publicationSite, publicBaseUrl, reqQuery, reqSession(req))
-  for (const r of response.results) {
-    datasetUtils.clean(req, r)
-  }
-  res.json(response)
-})
-
-router.use('/:datasetId/permissions', readDataset({ noCache: true }), apiKeyMiddlewareAdmin, rateLimiting.middleware, permissions.router('datasets', 'dataset', async (req, patchedDataset) => {
-  // this callback function is called when the resource becomes public
-  await publicationSites.onPublic(patchedDataset, 'datasets', reqSessionAuthenticated(req))
-}))
-
-// retrieve a dataset by its id
-router.get('/:datasetId', readDataset({ acceptInitialDraft: true, noCache: true }), apiKeyMiddlewareRead, rateLimiting.middleware, applicationKey, permissions.middleware('readDescription', 'read'), cacheHeaders.noCache, (req, res, next) => {
-  // @ts-ignore
-  const dataset = clone(req.dataset)
-  res.status(200).send(clean(req, dataset))
-})
-
-// retrieve only the schema.. Mostly useful for easy select fields
-const sendSchema = (req, res, schema) => {
-  schema = filterSchema(schema, req.query)
-  if (req.query.mimeType === 'application/tableschema+json') {
-    res.setHeader('content-disposition', contentDisposition(req.dataset.slug + '-tableschema.json'))
-    schema = tableSchema(schema)
-  } else if (req.query.mimeType === 'application/schema+json') {
-    res.setHeader('content-disposition', contentDisposition(req.dataset.slug + '-schema.json'))
-    schema = jsonSchema(schema, req.publicBaseUrl, req.query.arrays !== 'true')
-  } else {
-    for (const field of schema) {
-      field.label = field.title || field['x-originalName'] || field.key
-    }
-  }
-  res.status(200).send(schema)
-}
-router.get('/:datasetId/schema', readDataset(), apiKeyMiddlewareRead, rateLimiting.middleware, applicationKey, permissions.middleware('readSchema', 'read'), cacheHeaders.noCache, (req, res, next) => {
-  sendSchema(req, res, clone(req.dataset.schema))
-})
-// alternate read schema route that does not return clues about the data (cardinality and enums)
-router.get('/:datasetId/safe-schema', readDataset(), apiKeyMiddlewareRead, rateLimiting.middleware, applicationKey, permissions.middleware('readSafeSchema', 'read'), cacheHeaders.noCache, (req, res, next) => {
-  const schema = clone(req.dataset.schema)
-  for (const p of schema) {
-    delete p['x-cardinality']
-    delete p.enum
-  }
-  sendSchema(req, res, schema)
-})
-
-const permissionsWritePublications = permissions.middleware('writePublications', 'admin')
-const permissionsWriteExports = permissions.middleware('writeExports', 'admin')
-const permissionsSetReadApiKey = permissions.middleware('setReadApiKey', 'admin')
-const permissionsWriteDescription = permissions.middleware('writeDescription', 'write')
-const debugBreakingChanges = debugModule('breaking-changes')
-
-const descriptionBreakingKeys = ['rest', 'virtual', 'lineOwnership', 'primaryKey', 'projection', 'attachmentsAsImage', 'extensions', 'timeZone', 'slug'] // a change in these properties is considered a breaking change
-const descriptionHasBreakingChanges = (req) => {
-  const breakingChangeKey = descriptionBreakingKeys.find(key => key in req.body)
-  if (breakingChangeKey) {
-    debugBreakingChanges('breaking change on key', breakingChangeKey)
-    return true
-  }
-  if (!req.body.schema) return false
-  // TODO: some change in calculated properties should also be rejected here ?
-  const breakingChanges = getSchemaBreakingChanges(req.dataset.schema, req.body.schema, true, false)
-  debugBreakingChanges('breaking changes in schema ? ', breakingChanges)
-  return breakingChanges.length > 0
-}
-const permissionsWriteDescriptionBreaking = permissions.middleware('writeDescriptionBreaking', 'write')
-const permissionsManageMasterData = permissions.canDoForOwnerMiddleware('manageMasterData', true)
-
-// Update a dataset's metadata
-router.patch('/:datasetId',
-  readDataset({
-    acceptedStatuses: (patch, dataset) => {
-      // accept different statuses of the dataset depending on the content of the patch
-      if (!dataset.isMetaOnly && (patch.schema || patch.virtual || patch.extensions || patch.publications || patch.projection)) {
-        return ['finalized', 'error']
-      } else {
-        return null
-      }
-    }
-  }),
-  apiKeyMiddlewareWrite, rateLimiting.middleware,
-  lockDataset((/** @type {any} */ patch) => {
-    return !!(patch.schema || patch.virtual || patch.extensions || patch.publications || patch.projection)
-  }),
-  (req, res, next) => req.body.masterData ? permissionsManageMasterData(req, res, next) : next(),
-  (req, res, next) => {
-    if (descriptionHasBreakingChanges(req)) {
-      permissionsWriteDescriptionBreaking(req, res, next)
-    } else if (can('datasets', req.resource, 'writeDescription', reqSession(req))) {
-      permissionsWriteDescription(req, res, next)
-    } else {
-      // writeDescriptionBreaking implies writeDescription for non-breaking changes
-      permissionsWriteDescriptionBreaking(req, res, next)
-    }
-  },
-  (req, res, next) => req.body.publications ? permissionsWritePublications(req, res, next) : next(),
-  (req, res, next) => req.body.exports ? permissionsWriteExports(req, res, next) : next(),
-  (req, res, next) => req.body.readApiKey ? permissionsSetReadApiKey(req, res, next) : next(),
-  async (req, res) => {
-    // deep clone to allow mutation by applyPatch (req.dataset may be an immutable proxy from cache)
-    const dataset = clone(req.dataset)
-
-    const locale = req.getLocale()
-    const sessionState = reqSessionAuthenticated(req)
-
-    const { body: patch } = (await import('#doc/datasets/patch-req/index.js')).returnValid(req)
-
-    const { removedRestProps, attemptMappingUpdate, isEmpty } = await preparePatch(req.app, patch, dataset, sessionState, locale)
-
-    if (!isEmpty) {
-      await publicationSites.applyPatch(dataset, { ...dataset, ...patch }, sessionState, 'datasets')
-      await applyPatch(dataset, patch, removedRestProps, attemptMappingUpdate)
-        .catch(err => {
-          if (err.code !== 11000) throw err
-          throw httpError(400, req.__('errors.dupSlug'))
-        })
-
-      if (patch.status && patch.status !== 'indexed' && patch.status !== 'finalized' && patch.status !== 'validation-updated') {
-        await journals.log('datasets', dataset, { type: 'structure-updated' })
-        await notifications.sendResourceEvent('datasets', dataset, sessionState, 'structure-updated', { extra: { patch: Object.keys(patch).join(', ') } })
-      }
-
-      eventsLog.info('df.datasets.patch', `patched dataset ${dataset.slug} (${dataset.id}), keys=${JSON.stringify(Object.keys(patch))}`, { req, account: dataset.owner })
-
-      const draft = !!dataset.draftReason
-      eventsQueue.pushEvent({
-        title: `Propriétés modifiées sur un ${draft ? 'brouillon de ' : ''}jeu de données`,
-        body: `${draft ? 'brouillon ' : ''}${dataset.title} (${dataset.slug}), ${Object.keys(patch)?.join(', ')}`,
-        topic: {
-          key: `data-fair:dataset${draft ? '-draft' : ''}-patched-properties:${dataset.id}`
-        },
-        sender: dataset.owner,
-        resource: { type: 'dataset', title: dataset.title, id: dataset.id }
-      }, sessionState)
-
-      await syncRemoteService(dataset)
-    }
-
-    res.status(200).json(clean(req, dataset))
-  })
-
-// Change ownership of a dataset
-router.put('/:datasetId/owner', readDataset({ noCache: true }), apiKeyMiddlewareAdmin, rateLimiting.middleware, permissions.middleware('changeOwner', 'admin'), async (req, res) => {
-  // @ts-ignore
-  const dataset = req.dataset
-
-  const sessionState = reqSessionAuthenticated(req)
-
-  // Must be able to delete the current dataset, and to create a new one for the new owner to proceed
-  if (!sessionState.user.adminMode) {
-    if (req.body.type === 'user' && req.body.id !== sessionState.user.id) return res.status(403).type('text/plain').send(req.__('errors.missingPermission'))
-    if (req.body.type === 'organization') {
-      const userOrg = sessionState.user.organizations.find(o => o.id === req.body.id)
-      if (!userOrg) return res.status(403).type('text/plain').send(req.__('errors.missingPermission'))
-      if (![config.contribRole, config.adminRole].includes(userOrg.role)) return res.status(403).type('text/plain').send(req.__('errors.missingPermission'))
-    }
-  }
-
-  if (req.body.type !== req.dataset.owner.type || req.body.id !== req.dataset.owner.id) {
-    const remaining = await limits.remaining(req.body)
-    if (remaining.nbDatasets === 0) {
-      debugLimits('exceedLimitNbDatasets/changeOwner', { owner: req.body, remaining })
-      return res.status(429).type('text/plain').send(req.__('errors.exceedLimitNbDatasets'))
-    }
-    if (req.dataset.storage) {
-      if (remaining.storage !== -1 && remaining.storage < req.dataset.storage.size) {
-        debugLimits('exceedLimitStorage/changeOwner', { owner: req.body, remaining, storage: req.dataset.storage })
-        return res.status(429).type('text/plain').send(req.__('errors.exceedLimitStorage'))
-      }
-      if (remaining.indexed !== -1 && req.dataset.storage.indexed && remaining.indexed < req.dataset.storage.indexed.size) {
-        debugLimits('exceedLimitIndexed/changeOwner', { owner: req.body, remaining, storage: req.dataset.storage })
-        return res.status(429).type('text/plain').send(req.__('errors.exceedLimitIndexed'))
-      }
-    }
-  }
-
-  const patch = {
-    owner: req.body,
-    updatedBy: { id: sessionState.user.id, name: sessionState.user.name },
-    updatedAt: moment().toISOString()
-  }
-
-  const sameOrg = dataset.owner.type === 'organization' && dataset.owner.type === req.body.type && dataset.owner.id === req.body.id
-  if (sameOrg && !dataset.owner.department && req.body.department) {
-    // moving from org root to a department, we keep the publicationSites
-  } else {
-    patch.publicationSites = []
-  }
-  if (!sameOrg && req.body.publications) {
-    patch.publications = []
-  }
-
-  const preservePermissions = (dataset.permissions || []).filter(p => {
-    // keep public permissions
-    if (!p.type) return true
-    if (sameOrg) {
-      // keep individual user permissions (user partners)
-      if (p.type === 'user') return true
-      // keep permissions to external org (org partners)
-      if (p.type === 'organization' && p.id !== dataset.owner.id) return true
-    }
-    return false
-  })
-  await permissions.initResourcePermissions(patch, preservePermissions)
-
-  const patchedDataset = await mongo.db.collection('datasets')
-    .findOneAndUpdate({ id: req.dataset.id }, { $set: patch }, { returnDocument: 'after' })
-
-  // Move all files
-  if (dir(req.dataset) !== dir(patchedDataset)) {
-    try {
-      await filesStorage.moveDir(dir(req.dataset), dir(patchedDataset))
-    } catch (err) {
-      console.warn('Error while moving dataset directory', err)
-    }
-  }
-
-  const arrowStr = `${dataset.owner.name} (${dataset.owner.type}:${dataset.owner.id}) -> ${patch.owner.name} (${patch.owner.type}:${patch.owner.id})`
-  const eventLogMessage = `changed dataset owner ${dataset.slug} (${dataset.id}), ${arrowStr}`
-
-  eventsLog.info('df.datasets.changeOwnerFrom', eventLogMessage, { req, account: dataset.owner })
-  eventsLog.info('df.datasets.changeOwnerTo', eventLogMessage, { req, account: patch.owner })
-  const event = {
-    title: 'Changement de propriétaire d\'un jeu de données',
-    body: `${dataset.title} (${dataset.slug}), ${arrowStr}`,
-    topic: {
-      key: `data-fair:dataset-change-owner:${dataset.id}`
-    },
-    resource: { type: 'dataset', title: dataset.title, id: dataset.id },
-    sender: { ...dataset.owner, role: 'admin' }
-  }
-  eventsQueue.pushEvent(event, sessionState)
-  eventsQueue.pushEvent({ ...event, sender: { ...patch.owner, admin: true } }, sessionState)
-
-  await syncRemoteService(patchedDataset)
-
-  await updateTotalStorage(dataset.owner)
-  await updateTotalStorage(patch.owner)
-
-  res.status(200).json(clean(req, patchedDataset))
-})
-
-// Delete a dataset
-router.delete('/:datasetId', readDataset({ acceptedStatuses: ['*'], alwaysDraft: true }), apiKeyMiddlewareAdmin, rateLimiting.middleware, permissions.middleware('delete', 'admin'), async (req, res) => {
-  // @ts-ignore
-  const dataset = req.dataset
-  // @ts-ignore
-  const datasetFull = req.datasetFull
-
-  await deleteDataset(req.app, dataset)
-  if (dataset.draftReason && datasetFull.status !== 'draft') {
-    await deleteDataset(req.app, datasetFull)
-  }
-
-  eventsLog.info('df.datasets.delete', `dataset deleted ${dataset.slug} (${dataset.id})`, { req, account: dataset.owner })
-  const sessionState = await session.req(req)
-  eventsQueue.pushEvent({
-    title: 'Jeu de données supprimé',
-    body: `${dataset.title} (${dataset.slug})`,
-    topic: {
-      key: `data-fair:dataset-delete:${dataset.id}`
-    },
-    sender: dataset.owner
-  }, sessionState)
-
-  await syncRemoteService({ ...datasetFull, masterData: null })
-  await updateTotalStorage(datasetFull.owner)
-  res.sendStatus(204)
-})
+registerMetadataRoutes(router)
 
 const debugCreateDataset = debugModule('create-dataset')
